+++ Light.h
@@ -7,13 +7,13 @@
 
 class Light {
 
-    int m_ledPin;
+    uint8_t m_ledPin;
     bool m_isLedOn;
     Scheduler m_scheduler;
     CallbackTemplate<Light> m_togglePulseCallback;
 
 public:
-    Light(int pin);
+    Light(uint8_t pin);
 
     void on();
